+++ packages/core/src/Core.js
@@ -23,7 +23,6 @@ const {
 const tracer = require("./tracer");
 
 const Config = require("./Config");
-const Orchestrator = require("./Orchestrator");
 const Loaders = require("./Loaders");
 const Alerts = require("./Alerts");
 
@@ -38,6 +37,8 @@ const { scopedAlertsMethods, addEventListener } = require("./support/helpers");
 class Core {
   constructor(programmaticConfig) {
     this._eventEmitter = new EventEmitter();
+    this._loadedMocks = false;
+    this._loadedRoutes = false;
 
     this._alerts = new Alerts({
       onChange: (alerts) => {
@@ -48,12 +49,20 @@ class Core {
     this._mocksLoaders = new Loaders({
       onLoad: () => {
         this._eventEmitter.emit(LOAD_MOCKS);
+        this._loadedMocks = true;
+        if (this._loadedRoutes) {
+          this._mocks.load();
+        }
       },
     });
 
     this._routesLoaders = new Loaders({
       onLoad: () => {
         this._eventEmitter.emit(LOAD_ROUTES);
+        this._loadedRoutes = true;
+        if (this._loadedMocks) {
+          this._mocks.load();
+        }
       },
     });
 
@@ -66,7 +75,7 @@ class Core {
     this._settings = new Settings({
       onChange: (changeDetails) => {
         this._eventEmitter.emit(CHANGE_SETTINGS, changeDetails);
-        // TODO, define in options if they should produce a server restart
+        // TODO, define in options whether they should produce a server restart or not
         if (
           changeDetails.hasOwnProperty("port") ||
           changeDetails.hasOwnProperty("host") ||
@@ -116,17 +125,17 @@ class Core {
           this._alerts.rename
         ),
       },
-      this //To be used only by routeHandlers
+      this // To be used only by routeHandlers
     );
 
-    // TODO, refactor. Pass specific callbacks instead of objects
-    this._server = new Server(this._eventEmitter, this._settings, {
+    this._server = new Server({
+      getHostOption: () => this._settings.get("host"),
+      getPortOption: () => this._settings.get("port"),
+      getCorsOption: () => this._settings.get("cors"),
+      getCorsPreFlightOption: () => this._settings.get("corsPreFlight"),
       mocksRouter: this._mocks.router,
       ...scopedAlertsMethods("server", this._alerts.add, this._alerts.remove),
     });
-
-    // TODO, refactor. Pass specific callbacks instead of objects
-    this._orchestrator = new Orchestrator(this._eventEmitter, this._mocks);
 
     this._inited = false;
     this._stopPluginsPromise = null;